Finishing cardinalities in knowledge graph + adding detailed algorithm description which resolves #706 and resolves #696 and resolves #694 and resolves #693.

+++ app/src/qt/navigator/node.h
@@ -89,7 +89,7 @@ class NavigatorNode : public QGraphicsItem
     QString nodeType;
     QString nodeName;
     QString nodeCardinality;
-    int nodeCardinalityPixelWidth;
+    qreal nodeCardinalityPixelWidth;
     QColor nodeColor;
 
     bool showType;
